+++ addon/validators/minlength.js
@@ -1,0 +1,12 @@
+import Ember from 'ember';
+const { isEmpty, isNone } = Ember;
+
+export function minlength(value, minlength) {
+  return isEmpty(minlength) || isNone(value) || `${value}`.length >= parseInt(minlength, 10);
+}
+
+export default {
+  param: 'minlength',
+  message: 'Must have at least %@ characters.',
+  validate: minlength
+};