+++ tests/dummy/app/controllers/input.js
@@ -1,34 +1,46 @@
 import Ember from 'ember';
+const { Controller } = Ember;
 
-export default Ember.Controller.extend({
+export default Controller.extend({
 
-  multipleConstraints: [
-    {
-      'errorMessage': 'Value is not even',
-      'isError': (inputValue) => {
-        return (+inputValue % 2) === 1;
-      }
-    },
-    {
-      'errorMessage': 'Value does not equal 4',
-      'isError': (inputValue) => {
-        return +inputValue !== 4;
-      }
-    }
-  ],
+  multipleConstraints: [{
+    message: 'Value is not even',
+    validate: (inputValue) => (+inputValue % 2) === 0
+  }, {
+    message: 'Value does not equal 4',
+    validate: (inputValue) => +inputValue === 4
+  }],
 
-  singleContraint: {
-    'errorMessage': 'Value does not equal 16',
-    'isError': (inputValue) => {
-      return +inputValue !== 16;
-    }
-  },
+  singleContraint: [{
+    message: 'Value does not equal 16',
+    validate: (inputValue) => +inputValue === 16
+  }],
 
-  emailValidation: {
-    'errorMessage': 'Please provide email in a valid format',
-    'isError': (inputValue) => {
+  emailValidation: [{
+    message: 'Please provide email in a valid format',
+    validate: (inputValue) => {
       let emailPattern = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$/;
-      return !emailPattern.test(inputValue);
+      return emailPattern.test(inputValue);
     }
+  }],
+
+  eventName: '',
+  actions: {
+    focusReceived() {
+      this.set('eventName', 'focus');
+    },
+
+    blurReceived() {
+      this.set('eventName', 'blur');
+    },
+
+    keyDownReceived() {
+      this.set('eventName', 'keyDown');
+    },
+
+    enterReceived() {
+      this.set('eventName', 'enter');
+    }
+
   }
 });